Add WTZ TVL as new listing
