+++ projects/crunchynetwork/index.js
@@ -13,30 +13,11 @@ async function fetchFarmsTvl() {
 // crunchy farm v2 address KT1L1WZgdsfjEyP5T4ZCYVvN5vgzrNbu18kX
 // TVL = sum(crunchFarm.poolBalance / quipuLP.total_supply * quipuLP.tez_pool * 2 * XTZUSD)
 async function fetchFarmsV2Tvl() {
-    const farms = await get(RPC_ENDPOINT + '/v1/contracts/KT1KnuE87q1EKjPozJ5sRAjQA24FPsP57CE3/bigmaps/farms/keys?limit=1000')
+    const farms = await get(RPC_ENDPOINT + '/v1/contracts/KT1L1WZgdsfjEyP5T4ZCYVvN5vgzrNbu18kX/bigmaps/farms/keys?limit=1000')
     const items = farms.map(farm => [farm.value.poolToken.address, farm.value.poolBalance]).filter(item => item[1] !== "0")
     return getAllLPToTez(items);
 }
 
-async function fetchWtzTvl() {
-    const wtzContracts = [
-        "KT1BB3oNr5vUSw1CuPNb2zpYEVp376XrXWaJ",
-        "KT1H25LW5k4HQGm9hmNXzaxf3nqjsAEhQPah",
-        "KT1LpGZnT6dj6STSxHXmvSPqx39ZdPXAMpFz", 
-        "KT1NBgqqJacbdoeNAg9MvrgPT9h6q6AGWvFA",
-        "KT1NGTDBKDPMrAYEufb72CLwuQJ7jU7jL6jD"
-    ];
-
-    // Use Promise.all to fetch all wtz contract balances concurrently
-    const balances = await Promise.all(wtzContracts.map(async (contract) => {
-        const balance = await get(RPC_ENDPOINT + `/v1/accounts/${contract}/balance`);
-        return balance;
-    }));
-
-    const totalBalance = balances.reduce((total, balance) => total + balance, 0);
-    return totalBalance;
-
-}
 // crunchy freezer address KT1LjcQ4h5hCy9RcveFz9Pq8LtmF6oun7vNd
 // TVL = sum(cruchFreezer.amountLocked / quipuLP.total_supply * quipuLP.tez_pool * 2 * XTZUSD)
 async function fetchDeepFreezersTvl() {
@@ -77,9 +58,8 @@ async function tvl() {
     const farmsTvl = await fetchFarmsTvl();
     const farmsV2Tvl = await fetchFarmsV2Tvl();
     const deepFreezersTvl = await fetchDeepFreezersTvl();
-    const wtzTvl = await fetchWtzTvl();
     return {
-        tezos: (farmsTvl + farmsV2Tvl + deepFreezersTvl + wtzTvl ) / 1e6
+        tezos: (farmsTvl + farmsV2Tvl + deepFreezersTvl ) / 1e6
     };
 }
